+++ client/amqp/amqp_test.go
@@ -1,58 +1,34 @@
 package amqp
 
 import (
+	"context"
 	"testing"
 
-	"github.com/beatlabs/patron/examples"
+	"github.com/opentracing/opentracing-go"
+	"github.com/opentracing/opentracing-go/mocktracer"
+	"github.com/streadway/amqp"
 	"github.com/stretchr/testify/assert"
 )
 
-func TestNewMessage(t *testing.T) {
-	m := NewMessage("xxx", []byte("test"))
-	assert.Equal(t, "xxx", m.contentType)
-	assert.Equal(t, []byte("test"), m.body)
-}
-
-func TestNewJSONMessage(t *testing.T) {
-	m, err := NewJSONMessage("xxx")
-	assert.NoError(t, err)
-	assert.Equal(t, "application/json", m.contentType)
-	assert.Equal(t, []byte(`"xxx"`), m.body)
-	_, err = NewJSONMessage(make(chan bool))
-	assert.Error(t, err)
-}
-
-func TestNewProtobufMessage(t *testing.T) {
-	u := examples.User{
-		Firstname: "John",
-		Lastname:  "Doe",
-	}
-	m, err := NewProtobufMessage(&u)
-	assert.NoError(t, err)
-	assert.Equal(t, "application/x-protobuf", m.contentType)
-	assert.Len(t, m.body, 11)
-}
-
-func TestNewPublisher(t *testing.T) {
+func TestNew(t *testing.T) {
+	t.Parallel()
 	type args struct {
 		url string
-		exc string
-		opt OptionFunc
 	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
+	tests := map[string]struct {
+		args        args
+		expectedErr string
 	}{
-		{name: "fail, missing url", args: args{}, wantErr: true},
-		{name: "fail, missing exchange", args: args{url: "url"}, wantErr: true},
-		{name: "fail, missing exchange", args: args{url: "url", exc: "exc", opt: WithTimeout(0)}, wantErr: true},
+		"fail, missing url": {args: args{}, expectedErr: "url is required"},
 	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := NewPublisher(tt.args.url, tt.args.exc, tt.args.opt)
-			if tt.wantErr {
-				assert.Error(t, err)
+	for name, tt := range tests {
+		tt := tt
+		t.Run(name, func(t *testing.T) {
+			t.Parallel()
+
+			got, err := New(tt.args.url)
+			if tt.expectedErr != "" {
+				assert.EqualError(t, err, tt.expectedErr)
 				assert.Nil(t, got)
 			} else {
 				assert.NoError(t, err)
@@ -60,4 +36,14 @@ func TestNewPublisher(t *testing.T) {
 			}
 		})
 	}
+}
+
+func Test_injectTraceHeaders(t *testing.T) {
+	mtr := mocktracer.New()
+	opentracing.SetGlobalTracer(mtr)
+	t.Cleanup(func() { mtr.Reset() })
+	msg := amqp.Publishing{}
+	sp := injectTraceHeaders(context.Background(), "123", &msg)
+	assert.NotNil(t, sp)
+	assert.NotEmpty(t, msg.Headers)
 }
