+++ client/amqp/integration_test.go
@@ -1,7 +1,7 @@
 //go:build integration
 // +build integration
 
-package v2
+package amqp
 
 import (
 	"context"
